Use FileGuard for writing manifest
This will clean up the temporary manifest.new file if it cannot be
written fully. Although if that happens it is because of an IO error,
so it is likely that removing the file may fail due to the same
underlying cause. Still, in the case of a full disk this should help
a bit.
